requirements.txt
asgiref==3.2.10
Django==3.0.7
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
pytz==2020.1
sqlparse==0.3.1
